global catalog: perform installation from ipa-adtrust-install
