--- install/gc/tools/Makefile.am
@@ -1,7 +1,0 @@
-NULL =
-
-appdir = $(libexecdir)/ipa/gc/
-
-dist_app_SCRIPTS =		\
-	ipa-gc-install		\
-	$(NULL)
