+++ install/tools/ipa-adtrust-install.in
@@ -30,9 +30,10 @@ import sys
 import six
 
 from optparse import SUPPRESS_HELP  # pylint: disable=deprecated-module
+from optparse import OptionGroup  # pylint: disable=deprecated-module
 
 from ipalib.install import sysrestore
-from ipaserver.install import adtrust, service
+from ipaserver.install import adtrust, gc, service
 from ipaserver.install.installutils import (
     read_password,
     check_server_configuration,
@@ -91,6 +92,20 @@ def parse_options():
                       dest="enable_compat", default=False, action="store_true",
                       help="Enable support for trusted domains for old "
                            "clients")
+    gc_group = OptionGroup(parser, "global catalog options")
+    gc_group.add_option("--gc-password", sensitive=True, dest="gc_password",
+                        help="Directory Manager password for the Global "
+                             "Catalog")
+    gc_group.add_option("--gc-cert-file", dest="gc_cert_files",
+                        action="append", metavar="FILE",
+                        default=[],
+                        help="File containing the Global Catalog SSL "
+                             "certificate and private key")
+    gc_group.add_option("--gc-pin",  sensitive=True, dest="gc_pin",
+                        metavar="PIN",
+                        help="The password to unlock the Global Catalog "
+                             "private key")
+    parser.add_option_group(gc_group)
 
     options, _args = parser.parse_args()
     safe_options = parser.get_safe_opts(options)
@@ -146,6 +161,7 @@ def main():
     print("This includes:")
     print("  * Configure Samba")
     print("  * Add trust related objects to FreeIPA LDAP server")
+    print("  * Configure the Global Catalog")
     # TODO:
     # print "  * Add a SID to all users and Posix groups"
     print("")
@@ -210,7 +226,9 @@ def main():
             "Unrecognized error during check of admin rights: %s" % e)
 
     adtrust.install_check(True, options, api)
+    gc.install_check(api, options)
     adtrust.install(True, options, fstore, api)
+    gc.install(api, fstore, options)
 
     # Enable configured services and update DNS SRV records
     service.sync_services_state(api.env.host)
